Unify exchange contexts behind a trait
