+++ hylo-quotes/src/token_operation/stability_pool.rs
@@ -2,6 +2,7 @@
 
 use anyhow::{ensure, Context, Result};
 use fix::prelude::*;
+use hylo_core::exchange_context::ExchangeContext;
 use hylo_core::fee_controller::FeeExtract;
 use hylo_core::solana_clock::SolanaClock;
 use hylo_core::stability_pool_math::{
